update dble install shell for adapting first install

+++ behave_dble/compose/update_dble_version.sh
@@ -1,18 +1,33 @@
 #!/bin/bash
 # Copyright (C) 2016-2019 ActionTech.
 # License: https://www.mozilla.org/en-US/MPL/2.0 MPL version 2 or higher.
-#2.17.09.0
-#9.9.9.9
-#2.18.02.1
+#!/bin/bash
+set -e;
+#version like 2.19.09.0 or 9.9.9.9
 version=${1-"9.9.9.9"}
-conf_path="dble_`date '+%Y-%m-%d_%H:%M:%S'`"
+old_dble_exists=0
 
-/opt/dble/bin/dble stop
+if [ -x "/opt/dble/bin/dble" ]; then
+    old_dble_exists=1
+
+#   stop running dble
+    /opt/dble/bin/dble stop
+
+#   mv dble conf for reuse
+    mv /opt/dble/conf /tmp/${conf_path}
+
+    rm -rf /opt/dble
+fi
 rm -rf actiontech-dble.tar.gz
-cd /opt && wget ftp://ftp:ftp@10.186.18.20/actiontech-dble/qa/${version}/actiontech-dble.tar.gz \
-&& mv /opt/dble/conf /tmp/${conf_path}
-rm -rf /opt/dble
-cd /opt && tar -zxf actiontech-dble.tar.gz
-rm -rf /opt/dble/conf
-mv /tmp/${conf_path} /opt/dble/conf
-/opt/dble/bin/dble start
+cd /opt && wget ftp://ftp:ftp@10.186.18.20/actiontech-dble/qa/${version}/actiontech-dble.tar.gz
+tar -zxf actiontech-dble.tar.gz -C /opt
+
+if [ ${old_dble_exists} -eq 1 ]; then
+# reuse old dble conf if exists and start dble
+  conf_path="dble_`date '+%Y-%m-%d_%H:%M:%S'`"
+  rm -rf /opt/dble/conf \
+  && mv /tmp/${conf_path} /opt/dble/conf \
+  && /opt/dble/bin/dble start
+else
+  echo "dble is first installed!!!"
+fi
